refactor(clearly-defined): Bundle coordinate-related code
Move all model code related to coordinates to a new file for a better
overview.
