+++ clients/clearly-defined/src/main/kotlin/Curation.kt
@@ -94,50 +94,3 @@ data class Patch(
     val coordinates: Coordinates,
     val revisions: Map<String, Curation>
 )
-
-/**
- * See https://github.com/clearlydefined/service/blob/b339cb7/schemas/curations-1.0.json#L64-L83 and
- * https://docs.clearlydefined.io/using-data#a-note-on-definition-coordinates.
- */
-@Serializable(CoordinatesSerializer::class)
-data class Coordinates(
-    /**
-     * The type of the component. For example, npm, git, nuget, maven, etc. This talks about the shape of the
-     * component.
-     */
-    val type: ComponentType,
-
-    /**
-     * Where the component can be found. Examples include npmjs, mavencentral, github, nuget, etc.
-     */
-    val provider: Provider,
-
-    /**
-     * Many component systems have namespaces: GitHub orgs, NPM namespace, Maven group id, etc. This segment must be
-     * supplied. If your component does not have a namespace, use '-' (ASCII hyphen).
-     */
-    val namespace: String? = null,
-
-    /**
-     * The name of the component. Given the mentioned [namespace] segment, this is just the simple name.
-     */
-    val name: String,
-
-    /**
-     * Components typically have some differentiator like a version or commit id. Use that here. If this segment is
-     * omitted, the latest revision is used (if that makes sense for the provider).
-     */
-    val revision: String? = null
-) {
-    constructor(value: String) : this(value.split('/', limit = 5))
-
-    private constructor(parts: List<String>) : this(
-        type = ComponentType.fromString(parts[0]),
-        provider = Provider.fromString(parts[1]),
-        namespace = parts[2].takeUnless { it == "-" },
-        name = parts[3],
-        revision = parts.getOrNull(4)
-    )
-
-    override fun toString() = listOfNotNull(type, provider, namespace ?: "-", name, revision).joinToString("/")
-}
